+++ src/NeuralNet/Layers/Dropout.php
@@ -72,10 +72,15 @@ class Dropout implements Hidden
     /**
      * Return the width of the layer.
      *
-     * @return int|null
+     * @throws \RuntimeException
+     * @return int
      */
-    public function width() : ?int
+    public function width() : int
     {
+        if (!$this->width) {
+            throw new RuntimeException('Layer has not been initialized.');
+        }
+
         return $this->width;
     }